feat: 여러 게임을 동시에 진행할 수 있도록 상태패턴 제거
- Application 전역에서 공통으로 관리되던 turn 정보 게임 객체 안으로 이동
- 매번 요청 수신할 때 마다 DB에서 게임 로드한 뒤 처리하도록 수정
- 게임 진행 상황을 매번 자동으로 DB에 저장하도록 수정
- console application 삭제

+++ src/main/java/chess/dao/BoardDao.java
@@ -26,7 +26,7 @@ public class BoardDao {
         }
     }
 
-    public void update(int gameNo, Map<Position, Piece> board) {
+    public void update(long gameNo, Map<Position, Piece> board) {
         final String sql = "update piece set type = ?, white = ? where position = ? and game_no = ?";
 
         for (Entry<Position, Piece> entry : board.entrySet()) {
@@ -44,7 +44,7 @@ public class BoardDao {
         System.out.println("쿼리 실행 결과: " + result);
     }
 
-    public List<PieceDto> load(int gameNo) {
+    public List<PieceDto> load(long gameNo) {
         final String sql = "select type, white, position from piece where game_no = ?";
 
         return jdbcTemplate.query(sql, (resultSet, rowNum) -> PieceDto.of(
